ai-core/pyproject.toml
[project]
name = "ai-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate==1.11.0",
    "addict==2.4.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "attrs==25.4.0",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "datasets==3.6.0",
    "dill==0.3.8",
    "einops==0.8.1",
    "filelock==3.20.0",
    "frozenlist==1.8.0",
    "fsspec==2025.3.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "hf-xet==1.2.0",
    "huggingface-hub==0.36.0",
    "idna==3.11",
    "jinja2==3.1.6",
    "markupsafe==3.0.3",
    "modelscope==1.31.0",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "multiprocess==0.70.16",
    "networkx==3.5",
    "numpy==2.3.4",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-cupti-cu12==12.8.90",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.3.3.83",
    "nvidia-cufile-cu12==1.13.1.3",
    "nvidia-curand-cu12==10.3.9.90",
    "nvidia-cusolver-cu12==11.7.3.90",
    "nvidia-cusparse-cu12==12.5.8.93",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.5",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvshmem-cu12==3.3.20",
    "nvidia-nvtx-cu12==12.8.90",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==12.0.0",
    "propcache==0.4.1",
    "protobuf>=6.33.1",
    "psutil==7.1.3",
    "pyarrow==22.0.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "regex==2025.11.3",
    "requests==2.32.5",
    "safetensors==0.6.2",
    "scipy==1.16.3",
    "setuptools-scm==9.2.2",
    "simplejson==3.20.2",
    "six==1.17.0",
    "sortedcontainers==2.4.0",
    "sympy==1.14.0",
    "tokenizers==0.22.1",
    "torch==2.9.0",
    "torchaudio==2.9.0",
    "torchvision==0.24.0",
    "tqdm==4.67.1",
    "transformers==4.57.1",
    "triton==3.5.0",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "xxhash==3.6.0",
    "yarl==1.22.0",
]
